fix(anthropic): preserve unknown reasoning capability and reject conflicting effort
