+++ litellm/llms/anthropic/experimental_pass_through/messages/transformation.py
@@ -387,19 +387,20 @@ class AnthropicMessagesConfig(BaseAnthropicMessagesConfig):
                     ),
                     status_code=400,
                 )
-            gate_error: Final = AnthropicConfig._validate_effort_for_model(model, requested_effort, custom_llm_provider)
+            existing_output_config: Final = optional_params.get("output_config")
+            existing_mapping: Final = existing_output_config if isinstance(existing_output_config, dict) else {}
+            raw_explicit_effort: Final = existing_mapping.get("effort")
+            explicit_effort: Final = raw_explicit_effort if isinstance(raw_explicit_effort, str) else None
+            candidate_effort: Final = explicit_effort if explicit_effort is not None else requested_effort
+            gate_error: Final = AnthropicConfig._validate_effort_for_model(model, candidate_effort, custom_llm_provider)
             resolved_effort: Final = (
-                requested_effort
+                candidate_effort
                 if gate_error is None
-                else normalize_reasoning_effort_value(requested_effort, model, custom_llm_provider)
+                else normalize_reasoning_effort_value(candidate_effort, model, custom_llm_provider)
             )
-            if gate_error is not None and resolved_effort == requested_effort:
+            if gate_error is not None and resolved_effort == candidate_effort:
                 raise AnthropicError(message=gate_error, status_code=400)
-            existing_output_config = optional_params.get("output_config")
-            if not isinstance(existing_output_config, dict):
-                existing_output_config = {}
-            existing_output_config.setdefault("effort", resolved_effort)
-            optional_params["output_config"] = existing_output_config
+            optional_params["output_config"] = {**existing_mapping, "effort": resolved_effort}
 
     @staticmethod
     def _translate_adaptive_effort_for_non_adaptive_model(
