+++ tests/test_litellm/llms/anthropic/experimental_pass_through/messages/test_reasoning_effort_translation.py
@@ -1,5 +1,7 @@
 """Tests for ``reasoning_effort`` translation on the Anthropic /v1/messages route."""
 
+from unittest.mock import patch
+
 import pytest
 
 from litellm.constants import (
@@ -203,6 +205,43 @@ def test_reasoning_effort_max_accepted_on_sonnet_46_messages(
 
     output_config = result.get("output_config")
     assert isinstance(output_config, dict) and output_config.get("effort") == "max"
+
+
+def test_conflicting_unsupported_output_config_effort_is_not_forwarded():
+    with (
+        patch(
+            "litellm.llms.anthropic.common_utils.AnthropicModelInfo._is_adaptive_thinking_model",
+            return_value=True,
+        ),
+        patch(
+            "litellm.llms.anthropic.chat.transformation.AnthropicConfig._validate_effort_for_model",
+            side_effect=lambda model, effort, provider: (
+                None
+                if effort == "high"
+                else f"effort={effort!r} is not supported by this model. Got model: {model}"
+            ),
+        ),
+        patch(
+            "litellm.utils.get_model_info",
+            return_value={
+                "supports_reasoning": True,
+                "supports_max_reasoning_effort": False,
+                "supports_xhigh_reasoning_effort": False,
+            },
+        ),
+    ):
+        optional_params = {
+            "reasoning_effort": "high",
+            "output_config": {"effort": "max"},
+        }
+        AnthropicMessagesConfig._translate_reasoning_effort_to_anthropic(
+            model="claude-sonnet-4-6",
+            optional_params=optional_params,
+            max_tokens=1024,
+            custom_llm_provider="anthropic",
+        )
+        assert optional_params["output_config"]["effort"] != "max"
+        assert optional_params["output_config"]["effort"] == "high"
 
 
 def test_explicit_output_config_wins_over_reasoning_effort():
@@ -498,3 +537,58 @@ def test_disabled_thinking_omitted_for_always_on_models_messages(
         assert "thinking" not in result
     else:
         assert result["thinking"] == {"type": "disabled"}
+
+
+def _mock_model_info(**flags):
+    return flags
+
+
+def test_xhigh_degrades_to_high_for_non_adaptive_model():
+    with (
+        patch(
+            "litellm.llms.anthropic.common_utils.AnthropicModelInfo._is_adaptive_thinking_model",
+            return_value=False,
+        ),
+        patch(
+            "litellm.utils.get_model_info",
+            return_value=_mock_model_info(
+                supports_reasoning=True,
+                supports_xhigh_reasoning_effort=False,
+            ),
+        ),
+    ):
+        optional_params = {"reasoning_effort": "xhigh"}
+        AnthropicMessagesConfig._translate_reasoning_effort_to_anthropic(
+            model="unknown-glm-4.6",
+            optional_params=optional_params,
+            max_tokens=None,
+            custom_llm_provider="anthropic",
+        )
+        assert optional_params["thinking"]["type"] == "enabled"
+        assert optional_params["thinking"]["budget_tokens"] == DEFAULT_REASONING_EFFORT_HIGH_THINKING_BUDGET
+        assert "output_config" not in optional_params
+
+
+def test_max_degrades_to_high_for_non_adaptive_model():
+    with (
+        patch(
+            "litellm.llms.anthropic.common_utils.AnthropicModelInfo._is_adaptive_thinking_model",
+            return_value=False,
+        ),
+        patch(
+            "litellm.utils.get_model_info",
+            return_value=_mock_model_info(
+                supports_reasoning=True,
+                supports_max_reasoning_effort=False,
+                supports_xhigh_reasoning_effort=False,
+            ),
+        ),
+    ):
+        optional_params = {"reasoning_effort": "max"}
+        AnthropicMessagesConfig._translate_reasoning_effort_to_anthropic(
+            model="unknown-deepseek",
+            optional_params=optional_params,
+            max_tokens=None,
+            custom_llm_provider="anthropic",
+        )
+        assert optional_params["thinking"]["budget_tokens"] == DEFAULT_REASONING_EFFORT_HIGH_THINKING_BUDGET
